requirements.txt
Django==4.0
psycopg2-binary==2.9.3
Pillow==9.0
djangorestframework==3.13.1
django-filter==21.1
python-decouple==3.5
twilio==7.4
